// apps/frontend/src/__gen__/LinksFragment.graphql.ts
export type LinksFragment$data = {
  readonly links: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly actualUrl: string;
        readonly approved: boolean;
        readonly createdAt: any;
        readonly createdBy: string;
        readonly entity: {
          readonly name: string;
        } | null;
        readonly id: string;
        readonly name: string;
        readonly redirectUrl: string;
        readonly updatedAt: any;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: any | null;
      readonly hasNextPage: boolean | null;
      readonly hasPreviousPage: boolean | null;
      readonly startCursor: any | null;
    };
  };
  readonly " $fragmentType": "LinksFragment";
};